Update - Add food design complete
Custom Navigation View – This view alters the original look of the base navigational view to give it a new design.

Custom Pop up Window & Design – This is a view that appears above your layout and allows you to display a layout

Cool Transitions – I was able to figured out how to make views and design elements animate when used.

Nest layouts within layouts – I was able to use coordinator and recycle layouts to extend the capability's of my layouts